+++ reflectapi-demo/src/tests/serde.rs
@@ -692,3 +692,16 @@ fn test_datetime() {
 
     assert_snapshot!(TestStruct);
 }
+
+#[test]
+fn test_external_impls() {
+    #[derive(reflectapi::Input, reflectapi::Output, serde::Deserialize, serde::Serialize)]
+    struct Test {
+        index_map: reflectapi::indexmap::IndexMap<u8, u32>,
+        index_set: reflectapi::indexmap::IndexSet<String>,
+        url: reflectapi::url::Url,
+        json: reflectapi::serde_json::Value,
+    }
+
+    assert_snapshot!(Test);
+}
