+++ reflectapi/Cargo.toml
@@ -36,10 +36,10 @@ mime = { version = "0.3.17", optional = true }
 rmp-serde = {version = "1.3.0", optional = true }
 
 # optional 3rd party dependencies for implementing reflects traits for foreign types
-uuid = { version = "1.7.0", optional = true }
-chrono = { version = "0.4.37", optional = true }
-url = { version = "2.5.0", optional = true }
-rust_decimal = { version = "1.35.0", optional = true }
+uuid = { version = "1.7.0", optional = true, features = ["serde"] }
+chrono = { version = "0.4.37", optional = true, features = ["serde"] }
+url = { version = "2.5.0", optional = true, features = ["serde"] }
+rust_decimal = { version = "1.35.0", optional = true, features = ["serde"] }
 
 # optional 3rd party dependencies for transforming reflect schema
 # to runnable servers based on dififerent web server frameworks
@@ -48,7 +48,7 @@ axum = { version = "0.7.4", optional = true }
 # optional 3rd party dependencies for enabling codegen
 askama = { version = "0.12.1", optional = true }
 anyhow = { version = "1.0.81", optional = true }
-indexmap = { version = "2.2.6", optional = true }
+indexmap = { version = "2.2.6", optional = true, features = ["serde"] }
 check_keyword = { version = "0.2.0", optional = true }
 
 # optional 3rd party dependencies for client runtime
