+++ reflectapi/src/impls/mod.rs
@@ -7,21 +7,11 @@ mod chrono;
 #[cfg(feature = "url")]
 mod url;
 
+#[cfg(feature = "json")]
+mod json;
+
 #[cfg(feature = "rust_decimal")]
 mod rust_decimal;
 
-#[allow(unused_imports)]
-#[cfg(feature = "uuid")]
-pub use uuid::*;
-
-#[allow(unused_imports)]
-#[cfg(feature = "chrono")]
-pub use chrono::*;
-
-#[allow(unused_imports)]
-#[cfg(feature = "url")]
-pub use url::*;
-
-#[allow(unused_imports)]
-#[cfg(feature = "rust_decimal")]
-pub use rust_decimal::*;
+#[cfg(feature = "indexmap")]
+mod indexmap;
